리눅스 빌드 warning 을 제거함

+++ SipPlatform/TimeString.cpp
@@ -203,7 +203,7 @@ time_t ParseCookieExpires( const char * pszTime )
 		}
 		else if( cType )
 		{
-			if( iBufPos >= ( sizeof(szBuf) - 1 ) ) return 0;
+			if( iBufPos >= ( (int)sizeof(szBuf) - 1 ) ) return 0;
 			szBuf[iBufPos++] = pszTime[i];
 		}
